Create data.js and add objects
