+++ index.html
@@ -121,89 +121,6 @@
       <div class="under-bar"></div>
 
       <div class="works">
-        <div class="work">
-          <img src="images/default.png" alt="project overwiew image" />
-          <div class="work-text">
-            <h2 class="work-description">Multi-Post Stories Gain+Glory</h2>
-            <ul class="used-langages">
-              <li>Ruby on rails</li>
-              <li>css</li>
-              <li>Javascript</li>
-              <li>html</li>
-            </ul>
-            <div class="work-button">See Project</div>
-          </div>
-        </div>
-
-        <div class="work">
-          <img src="images/default.png" alt="project overwiew image" />
-          <div class="work-text">
-            <h2 class="work-description">Multi-Post Stories Gain+Glory</h2>
-            <ul class="used-langages">
-              <li>Ruby on rails</li>
-              <li>css</li>
-              <li>Javascript</li>
-              <li>html</li>
-            </ul>
-            <div class="work-button">See Project</div>
-          </div>
-        </div>
-
-        <div class="work">
-          <img src="images/default.png" alt="project overwiew image" />
-          <div class="work-text">
-            <h2 class="work-description">Multi-Post Stories Gain+Glory</h2>
-            <ul class="used-langages">
-              <li>Ruby on rails</li>
-              <li>css</li>
-              <li>Javascript</li>
-              <li>html</li>
-            </ul>
-            <div class="work-button">See Project</div>
-          </div>
-        </div>
-
-        <div class="work">
-          <img src="images/default.png" alt="project overwiew image" />
-          <div class="work-text">
-            <h2 class="work-description">Multi-Post Stories Gain+Glory</h2>
-            <ul class="used-langages">
-              <li>Ruby on rails</li>
-              <li>css</li>
-              <li>Javascript</li>
-              <li>html</li>
-            </ul>
-            <div class="work-button">See Project</div>
-          </div>
-        </div>
-
-        <div class="work">
-          <img src="images/default.png" alt="project overwiew image" />
-          <div class="work-text">
-            <h2 class="work-description">Multi-Post Stories Gain+Glory</h2>
-            <ul class="used-langages">
-              <li>Ruby on rails</li>
-              <li>css</li>
-              <li>Javascript</li>
-              <li>html</li>
-            </ul>
-            <div class="work-button">See Project</div>
-          </div>
-        </div>
-
-        <div class="work">
-          <img src="images/default.png" alt="project overwiew image" />
-          <div class="work-text">
-            <h2 class="work-description">Multi-Post Stories Gain+Glory</h2>
-            <ul class="used-langages">
-              <li>Ruby on rails</li>
-              <li>css</li>
-              <li>Javascript</li>
-              <li>html</li>
-            </ul>
-            <div class="work-button">See Project</div>
-          </div>
-        </div>
       </div>
     </section>
 
@@ -299,6 +216,6 @@ Typing something </textarea
       <h3>myemailhere@email.com</h3>
       <div class="end-bar"></div>
     </footer>
-    <script src="script.js"></script>
+    <script type="module"src="source/script.js"></script>
   </body>
 </html>
