+++ source/data.js
@@ -1,0 +1,39 @@
+let projet1 ={
+    imageSrc: "images/default.png",
+    titre: "Multi-Post Stories Gain+Glory",
+    technolgies: ["Ruby on rails","css","javascript",'html'],
+    link: "www.nicole.com"
+  },
+   projet2 ={
+    imageSrc: "images/default.png",
+    titre: "Multi-Post Stories Gain+Glory",
+    technolgies: ["Ruby on rails","css","javascript",'html'],
+    link: "www.nicole.com"
+  },
+   projet3 ={
+    imageSrc: "images/default.png",
+    titre: "Multi-Post Stories Gain+Glory",
+    technolgies: ["Ruby on rails","css","javascript",'html'],
+    link: "www.nicole.com"
+  },
+  projet4 ={
+    imageSrc: "images/default.png",
+    titre: "Multi-Post Stories Gain+Glory",
+    technolgies: ["Ruby on rails","css","javascript",'html'],
+    link: "www.nicole.com"
+  },
+  projet5 ={
+    imageSrc: "images/default.png",
+    titre: "Multi-Post Stories Gain+Glory",
+    technolgies: ["Ruby on rails","css","javascript",'html'],
+    link: "www.nicole.com"
+  },
+  projet6 ={
+    imageSrc: "images/default.png",
+    titre: "Multi-Post Stories Gain+Glory",
+    technolgies: ["Ruby on rails","css","javascript",'html'],
+    link: "www.nicole.com"
+  };
+  let works = [projet1,projet2,projet3,projet4,projet5,projet6];
+
+ export {works};
